netlify/functions/send-thank-you/requirements.txt
google-generativeai
# You would add 'sendgrid' here if you implement email sending.